tapfreighter: move coin selection code to own file
