+++ tapfreighter/coin_select_test.go
